server_backend/requirements.txt
paho-mqtt==1.6.1
pandas
scikit-learn
joblib